w3c-reports/12092014/efl/runned-tests/websockets/Secure-Send-data-expected.txt

CONSOLE MESSAGE: line 78: Error: wss not yet supported

Harness Error. harness_status.status = 1 , harness_status.message = Error: wss not yet supported
